Manual

Manual

Reading and writing

Reading and searching shapefile

Reading and writing list file

Reading and writing raster layers

Reading and writing Mikrubi field

Reading and writing Mikrubi model

Rasterizing a shapefile

Processing the raster layers

The Mikrubi core

Mikrubi field

Mikrubi model

Fitting a Mikrubi model

Predicting from a Mikrubi model

Sampling counties in a Mikrubi field

Mathematic functions

Reading and writing

Reading and searching shapefile

Since Mikrubi always reads instead of writes shapefile, only the reading function is offered.

Mikrubi.readshape — Function

readshape(path::AbstractString)

Reads a shape (".shp") file located at path. Alias for Shapefile.Table.

The function lookup is useful when some attribute (e.g. name or code) of a county is known and the row number of the county in a shapefile is wanted (row number acts as identifiers in the list of occupied counties, see the syntax of fit).

Mikrubi.lookup — Function

lookup(shptable::Shapefile.Table, column::Symbol, entry)
lookup(shptable::Shapefile.Table, column::Symbol, entries::AbstractArray)

Finds row(s) in the shapefile table whose column record(s) equal(s) to entry or elements of entries. When the third argument is an array, results are output as an array of the same shape by broadcasting.

Reading and writing list file

List of occupied counties can be prepared explicitly in REPL as a vector or a set. Meanwhile, it is also possible to read from or write to disk such a list, especially when the list is generated outside Julia.

Mikrubi.readlist — Function

readlist(path::AbstractString)

Reads any list of vector from file.

Mikrubi.writelist — Function

writemodel(path::AbstractString, list::Vector)

Writes any list or vector to file.

Reading and writing raster layers

Climatic factors are downloaded and stored as raster layers. Mikrubi reads such layers by readlayers, performs principal component analysis on them and returns the results as layers also. When the output layers need to be kept for future use, they can be written to disk using writelayers. Moreover, when the predicted distribution of species is organized in raster format, it can be saved likewise using writelayer.

Mikrubi.readlayers — Function

readlayers(dir::AbstractString; extset=nothing)

Returns all raster layers as a vector of GeoArrays from the directory dir. All elements in the vector returned are modifiable, and modifications do not affect the files on the disk.

By setting extset to a container of extensions (e.g. Set(".tif"), or [".tiff"]), only files ending with its elements are read. Default value is nothing, which means no file extension filtering.

Mikrubi.writelayer — Function

writelayer(path::AbstractString, layer::GeoArray)

Writes layer to the disk at path. Alias for GeoArrays.write!.

Mikrubi.writelayers — Function

writelayers(paths::Vector{<:AbstractString}, layers::Vector)
writelayers(pathformula::AbstractString, layers::Vector)

Writes layers to paths respondingly, or a series of paths generated by the pathformula where an asterisk is used for wildcard and replaced by numbers.

Internal functions

It worths mention that when reading layers from a directory, files are sorted according to their names in a manner similar to the sorting order in Windows OS. Please pay extra attention when two parallel series of raster layers are involved for makefield.

Mikrubi.sortfilenames! — Function

sortfilenames!(filenames::Vector{<:AbstractString})

Identifies the distinctive parts among filenames, and sorts filenames according to the order of those parts. If all of the distinctive parts are decimal numerals, they are sorted as integers.

Examples

julia> sortfilenames!(["bio_9.tif", "bio_10.tif", "bio_1.tif"])
[ Info: Totally 3 raster files recognized in the directory, in the form of bio_*.tif, where asterisk mean 1, 9, 10.
3-element Array{String,1}:
 "bio_1.tif"
 "bio_9.tif"
 "bio_10.tif"

julia> sortfilenames!(["bio_09.tif", "bio_10.tif", "bio_01.tif"])
[ Info: Totally 3 raster files recognized in the directory, in the form of bio_*.tif, where asterisk mean 01, 09, 10.
3-element Array{String,1}:
 "bio_01.tif"
 "bio_09.tif"
 "bio_10.tif"

All layers read are copied (so that they are no longer read-only and linking to the files on disk) and replaced extreme values by missings. To change this behavior, please refer to the inner function Mikrubi.copylayer.

Mikrubi.copylayer — Function

copylayer(layer::GeoArray; 
	fixmissing=true, verylargebase=1_000_000_000_000)

Copies a GeoArray layer (so that the matrix is not read-only).

By setting fixmissing = true, all pixels with absolute values greater than verylargebase (= 10^12 by default) are identified as missing.

Reading and writing Mikrubi field

MikrubiField is a specially designed type where the environmental information of pixels and their county identifiers are nested. It may be necessary to save (by writefield) and load (by readfield) a Mikrubi field especially when it is used on multiple species.

Mikrubi.readfield — Function

readfield(path::AbstractString)

Reads a Mikrubi field from file.

Mikrubi.writefield — Function

writefield(path::AbstractString, field::MikrubiField)

Writes a Mikrubi field to file.

Reading and writing Mikrubi model

MikrubiModel is a struct containing transformation parameters. It can be read from and written to disk using respectively readmodel and writemodel.

Mikrubi.readmodel — Function

readmodel(path::AbstractString)

Reads a Mikrubi model from file.

Mikrubi.writemodel — Function

writemodel(path::AbstractString, model::MikrubiModel)

Writes a Mikrubi model to file.

Rasterizing a shapefile

Rasterization of one or multiple polygon(s) in a shapefile is implemented in the function rasterize. The implementation is based on such a fact, that a pixel is touched by a polygon (simple or self-intersecting) when and only when either some edge of the polygon runs through the pixel or the center of the pixel lies inside the polygon. Pixels satisfying the two conditions are collected respectively by internal functions Mikrubi.scanline and Mikrubi.strokepath.

Mikrubi.rasterize — Function

rasterize(geom::Shapefile.Polygon, grid::GeoArray)

Returns coordinates of pixels of grid that geom the polygon(s) touch(es).

rasterize(shptable::Shapefile.Table, grid::GeoArray)

Collects coordinates of pixels of grid that every polygon(s) from shptable touch(es).

When rasterize is applied to an entire shape table (containing multiple counties), its returned value is of an internal type Mikrubi.CtPixels.

Mikrubi.CtPixels — Type

CtPixels

Alias for Vector{Tuple{Tuple{Int, Int}, Int}}, each element formatted as ((i, j), ctid), where i and j are matrix index on row and column, and ctid is identifier of a county.

Internal functions

These internal functions are called directly or indirectly inside rasterize.

Mikrubi.interpolate — Function

interpolate(u1::Float64, u2::Float64, v1::Float64, v2::Float64)
interpolate(u1, u2, v1, v2)

Returns an affine function as a closure, say f, which satisfies f(u1) == v1 and f(u2) == v2.

WARNING: Arguments u1 and u2 cannot be identical. Beforehand judgement is required.

Mikrubi.scanline — Function

scanline(coords::Matrix{Float64}, parts::Vector{<:Integer}, xb, yb)

Performs scan line algorithm on polygon(s) defined by coords and parts, in the rectangular field ranging from 0 to xb on longitude and from 0 to yb on latitude, with grid resolution of 1 unit. Argument coords is a matrix containing two rows representing longitude and latitude respectively, and parts is a vector of integers denoting starting column index of each ring in the polygon(s). When the center of a pixel lies within the polygon(s), the center coordinates of the pixel are included in the output.

Mikrubi.trimedge — Function

trimedge(x1, y1, x2, y2, xb, yb)

Trims the segment connecting (x1, y1) and (x2, y2) by a rectangle ranging from 0 to xb on longitude and from 0 and yb on latitude.

Example

julia> trimedge(-2.0, -3.0, 10.0, 6.0, 6.0, 7.0)
(2.0, 0.0, 6.0, 3.0)

Mikrubi.strokeedge — Function

strokeedge(x1, y1, x2, y2, xb, yb)

Gathers all unit pixels in the rectangle ranging from 0 to xb on longitude and from 0 and yb on latitude that the segment connecting (x1, y1) and (x2, y2) touches.

Mikrubi.strokepath — Function

strokepath(coords::Matrix{Float64}, parts, xb, yb)

Gathers all unit pixels in the rectangle ranging from 0 to xb on longitude and from 0 and yb on latitude that the path(s) defined by coords and parts touch(es).

Mikrubi.geom2mat — Function

geom2mat(geom::Shapefile.Polygon)

Converts a Shapefile.Polygon to a two-row matrix, with first row representing the longitudes, and the second for latitudes.

Processing the raster layers

In Mikrubi, climatic factors after being read in typically undergo some processing steps together with shapefile inside the function makefield, which returns a Mikrubi field and a stack of extracted components in raster layers. The two outputs can be used for training and prediction.

Sometimes it is also required to apply a model to another circumstance (different time or different space), in which case another series of parallel climatic factor layers need to be processed in exactly the same way as those used to generate the Mikrubi field (so that their climatic meanings are the same). Such layers need to be put in the third place in the input argument list for makefield, and those parallelly extracted components are returned in the third place in output as well.

Mikrubi.makefield — Function

makefield(layers::Vector{<:GeoArray}, shptable::Shapefile.Table; 
	rabsthres=0.8, nprincomp=3)
makefield(layers::Vector{<:GeoArray}, shptable::Shapefile.Table,
	players::Vector{<:GeoArray}; rabsthres=0.8, nprincomp=3)
makefield(layers::Vector{<:GeoArray}, ctpixels::CtPixels; 
	rabsthres=0.8, nprincomp=3)
makefield(layers::Vector{<:GeoArray}, ctpixels::CtPixels,
	players::Vector{<:GeoArray}; rabsthres=0.8, nprincomp=3)

Generates a Mikrubi field as well as processed variable layers from layers and shptable or ctpixels, by

(rasterizing the shptable to ctpixels using rasterize,)

masking the layers with ctpixels (using Mikrubi.masklayers!),

extracting non-missing pixels from layers (using Mikrubi.extractlayers),

selecting less correlated variables (using Mikrubi.selectvars), and

doing the principal component analysis (using Mikrubi.princompvars).

Optional keyword arguments

rabsthres: threshold of collinearity.

Absolute value of Pearson correlation efficient greater than this threshold is identified as collinearity and the two variables are thus incompatible.

nprincomp: expected number of principal components of the variables.

Notes about players

when players is present in the argument list, raster layers it contains experience the same process including subsetting, selecting, and taking principal components, and results are packed and returned in the third place. Users must assure that players has the same length as layers, and their elements are corresponding in order. This would be useful when the prediction is in another geographic range or at another time.

Internal functions

These internal functions are called directly or indirectly inside makefield. Customization is possible when necessary, for example, when the affine transformation during principal component analysis is explicitly wanted.

Mikrubi.masklayers! — Function

masklayers!(layers::Vector{<:GeoArray}, ctpixels::CtPixels)

Masks the layers in a way that only pixels present in ctpixels are kepts, while all other unmentioned pixels are set to missing.

Mikrubi.extractlayers — Function

extractlayers(layers::Vector{<:GeoArray})

Extracts the non-missing pixels from the layers, and combines them into a matrix, whose rows representing pixels and columns representing variables.

extractlayers is the inverse function of makelayers.

Mikrubi.emptylayer — Function

emptylayer(grid::GeoArray)

Creates a GeoArray with the same size and other attributes as grid, but full of missings in its matrix.

Mikrubi.makelayer — Function

makelayer(vector::Vector, idx::Vector, grid::GeoArray)

Returns a layer according to the grid from grid and content values from vector. See also makelayers.

Mikrubi.makelayers — Function

makelayers(matrix::Matrix, idx::Vector, grid::GeoArray)

Returns a vector of layers according to the grid from grid and content values from columns of matrix. See also makelayer.

makelayers is the inverse function of extractlayers.

Mikrubi.dftraverse! — Function

dftraverse!(beststate, bestscore, state, score, depth, maxdepth, 
	incompat, scoremat)

Finds the index combination that

firstly containing as many indices as possible, and

secondly with the lowest pairwise sum from submatrix of scoremat,

such that no indices i and j coexist as long as incompat[i][j] == true.

The result is stored as the only element of beststate, with its score decided by the two criteria above stored as the only element of bestscore.

Example

julia> beststate = Vector(undef, 1);

julia> bestscore = [(0, 0.0)];

julia> dftraverse!(beststate, bestscore, Int[], (0, 0.0), 1, 3,
           Bool[0 0 1; 0 0 0; 1 0 0],
           [0.0 0.6 0.3; 0.6 0.0 0.9; 0.3 0.9 0.0]);

julia> beststate
1-element Array{Any,1}:
 [1, 2]

julia> bestscore
1-element Array{Tuple{Int64,Float64},1}:
 (2, -0.6)

Mikrubi.selectvars — Function

selectvars(matrix::Matrix, rabsthres=0.8)

Selects as many variables as possible from matrix such that no pairwise Pearson coefficient among them exceeds rabsthres and their sum is minimal.

Example

julia> selectvars([1. 4. 7.; 2. 5. 8.; 3. 9. 27.], rabsthres=0.9)
2-element Array{Int64,1}:
 1
 3

Mikrubi.princompvars — Function

princompvars(submat::Matrix; nprincomp=3)

Performs principal component analysis on submat whose columns represents variables, and combines the nprincomp principal components into a matrix, and returns the result matrix as well as the affine transformation (colmean, projwstd), such that the result matrix == (submat .- colmean) * projwstd.

The Mikrubi core

Two specially designed structs are involved in the core of Mikrubi.

Mikrubi field

MikrubiField is a struct containing mainly three types of information of pixels/points, that is, which counties they belong to (ctids), their geographic coordinates (locs), and their environmental coordinates (vars), with also some derived assistant attributes, such as geographic dimensionality (usually 2) and environmental dimensionality (for example, 3).

MikrubiField can be obtained in three ways: as first output argument of makefield, read from disk, or constructed directly from the three required attributes (this may be useful for simulation analysis).

Mikrubi.MikrubiField — Type

MikrubiField(ctids, locs, vars)

Constructs a Mikrubi field containing a number of pixels or points, using the arguments which should always have the same number of rows

ctids::Vector: a vector containing the county identifiers

locs::Array{<:Real}: an array of geographic coordinates

vars::Matrix{<:AbstractFloat}: an array of environmental coordinates

Mikrubi model

MikrubiModel contains the environmental dimensionality and the model parameters to define a positive-definite quadratic mapping from environmental space to a real number axis.

Like MikrubiField, MikrubiField can be obtained in three ways: as output argument of fit, read from disk, or constructed directly from attributes. An example of obtaining Mikrubi field and Mikrubi model from constructors are available in examples/onedimsim/sim.jl.

Mikrubi.MikrubiModel — Type

MikrubiModel(dvar::Int, params::Vector{<:AbstractFloat})

Constructs a Mikrubi Model from a dimensionality dvar and a parameter vector params. The equation must hold for dvar2dparam(dvar) == length(params).

Mikrubi Models can be obtained from the function fit, and can be used in the function predict.

Fitting a Mikrubi model

When a Mikrubi field as well as occurrence data in county and/or in coordinates are ready, they can be used to train a Mikrubi model by function fit (county data at counties, required; coordinates at coords, optional). Result is output as a MikrubiModel.

StatsBase.fit — Function

fit(field::MikrubiField, counties, coords=zeros(0, 0); 
	optresult=[], iterations=3_000_000, kwargs...)

Numerically finds the Mikrubi model maximizing the likelihood that the occupied counties as well as the occupied coordinates are sampled in the given Mikrubi field. The optimization result is stored in the container optresult for debugging.

Predicting from a Mikrubi model

A Mikrubi model can be applied by function predict to a matrix with its columns corresponding to extracted variables, a stack of extracted layers, or a Mikrubi field.

When input argument is a matrix, output argument is a column vector denoting the probability of presence in pixels/points related to rows in the matrix.

When input argument is a stack of layers, output argument is a single layer denoting the probability of presence.

When input argument is a Mikrubi field, output argument is a Dict which maps every county identifier to probability of presence at pixels inside the county, see also predictcounty.

Mikrubi.predict — Function

predict(matrix::Matrix, model::MikrubiModel)
predict(layers::Vector{<:GeoArray}, model::MikrubiModel)
predict(field::MikrubiField, model::MikrubiModel)

Predicts the probability of presence according to processed climatic factors (matrix / layers) or on the Mikrubi field.

When distribution probability within only one county is concerned, predictcounty returns probability of presence at all pixels that constitute the county in descending order. Therefore, the first element represents the most likely occupied pixel of a county.

Mikrubi.predictcounty — Function

predictcounty(field::MikrubiField, model::MikrubiModel, county)

Returns the geographic coordinates of pixels in the county sorted by the likeliness of being occupied.

It is also possible to obtain the overall probability that every county is occupied by the function probcounties.

Mikrubi.probcounties — Function

probcounties(field::MikrubiField{T, U, V}, model::MikrubiModel{V})

Returns the probability of being occupied for every county in the field.

Sampling counties in a Mikrubi field

For simulation analysis, sometimes it is required to sample a set of counties from a Mikrubi field and a Mikrubi model. samplecounties does the trick.

Mikrubi.samplecounties — Function

samplecounties(field::MikrubiField, model::MikrubiModel)

Samples the counties according to their probability of being occupied.

Mathematic functions

Two mathematic functions used in the core are exported for convenience.

Mikrubi.logistic — Function

logistic(x)

Computes logistic(x) := 1 / (1 + e^x).

Mikrubi.loglogistic — Function

logistic(x)

Computes log(logistic(x)) = -log(1 + e^x).

Internal functions

Mikrubi.dvar2dparam — Function

dvar2dparam(dvar::Int)

Transforms dimensionality of an environmental space to the dimensionality of the induced parameter space, i.e., returns the degrees of freedom for positive-definite quadratic functions mapping a dvar-dimensional linear space into real numbers.

Examples

julia> dvar2dparam(1)
3

julia> dvar2dparam(3)
10

Mikrubi.decomparams — Function

decomparams(p::Vector, d::Int)
decomparams(model::MikrubiModel)

Returns parameter decomposition A, b, c, where

A is an upper triangular matrix of size (d, d),

b is a column vector of size d, and

c is a scalar.

WARNING: The vector p must have length dvar2dparam(d).

Example

julia> decomparams(collect(1:10), 3)
([1 0 0; 2 3 0; 4 5 6], [7, 8, 9], 10)

Mikrubi.loglike — Function

loglike(field::MikrubiField, params::Vector)

Computes the log-likelihood of being absent in each pixel given field and params.

Mikrubi.energy — Function

energy(field::MikrubiField, counties, params::Vector)
energy(vars::AbstractMatrix, params::Vector)

Computes the opposite log-likelihood that the occupied counties or occupied coordinates are sampled. The result is taken opposite sign for optimization, and therefore the function is called energy.

Mikrubi.probpixels — Function

probpixels(field::MikrubiField, model::MikrubiModel)

Returns the probability of being occupied for every pixel in the field.

Mikrubi.findnearest — Function

findnearest(loc::AbstractVecOrMat{<:Real}, field::MikrubiField)

Returns the row number in field.locs which is the nearest to the given coordinates.

Mikrubi.findnearests — Function

findnearests(loc::Vector{<:AbstractVecOrMat}, field::MikrubiField)
findnearests(loc::Matrix{<:Real}, field::MikrubiField)

Returns the row numbers in field.locs which are the nearest to each of the given coordinates. Duplicate results are reduced to one.